// src/app/constants/API.enum.ts
export enum APIEnum {
  REGISTRATION = "http://localhost:3000/auth",
  LOGIN = "http://localhost:3000/auth/login",
  AUTH_GET = "http://localhost:3000/auth",
  GET_ALL_PLANES = "http://localhost:3000/plan",
  GET_PLAN_BY_ID = "http://localhost:3000/plan",
  UPDATE_CUSTOMER_INFO = "http://localhost:3000/customer",
  GET_CUSTOMER_BY_ID = "http://localhost:3000/customer",
  GET_ALL_CUSTOMERS = "http://localhost:3000/customer",
  CREATE_ORDER = "http://localhost:3000/order",
  GET_ORDER_BY_CUSTOMER = "http://localhost:3000/order",
  GET_ALL_ORDERS = "http://localhost:3000/order",
  PRODUCTS = "http://localhost:3000/product",
  CATEGORIES = "http://localhost:3000/product/category",
  NOTIFICATIONS = "http://localhost:3000/note",
  CHAT = "http://localhost:3000/chat",
  CUSTOMER_PHONE_NUMBER = "http://localhost:3000/customer/search",
  CALL = "http://localhost:3000/call"
}